+++ src/enums/platform-type.ts
@@ -1,0 +1,7 @@
+export enum PlatformType {
+    Windows = 'windows',
+    Apple = 'apple',
+    Android = 'android',
+    Linux = 'linux',
+    Web = 'web',
+}